- Added active client/admin - Added Product (view add del) - Added SubProduct (view del)

+++ src/app/products/products.component.ts
@@ -6,6 +6,9 @@ import {ProductService} from '../../services/product.service';
 import {Product} from '../../model/product.model';
 import {Subproduct} from '../../model/subproduct.model';
 import {SubproductService} from '../../services/subproduct.service';
+import {ClientService} from '../../services/client.service';
+
+declare var $: any;
 
 @Component({
     selector: 'app-products',
@@ -22,19 +25,49 @@ export class ProductsComponent implements OnInit {
     subProductsPage: any;
     subProductDataSource = new MatTableDataSource(this.subProductsPage);
 
-    @ViewChild(MatPaginator) paginator: MatPaginator;
-    @ViewChild(MatSort) sort: MatSort;
+    @ViewChild(MatPaginator) productPaginator: MatPaginator;
+    @ViewChild('productSort') productSort: MatSort;
 
-    constructor(public http: HttpClient, public router: Router, public productservice: ProductService, public subproductservice: SubproductService) {
+    @ViewChild(MatPaginator) subProductPaginator: MatPaginator;
+    @ViewChild('subProductSort') subProductSort: MatSort;
+
+    constructor(public http: HttpClient, public router: Router, public productservice: ProductService, public subproductservice: SubproductService, public clientservice: ClientService) {
 
     }
 
-    ngAfterViewInit() {
-        this.productDataSource.paginator = this.paginator;
-        this.productDataSource.sort = this.sort;
+    showNotification(from, align, color, message) {
+        const type = ['danger', 'success'];
+        const icon = ['error_outline', 'check'];
 
-        // this.productsPage.paginator = this.paginator;
-        // this.productsPage.productSort = this.productSort;
+        $.notify(
+            {
+                message: message
+            },
+            {
+                type: type[color], // {0}
+                delay: 2000,
+                placement: {
+                    from: from,
+                    align: align
+                },
+                template:
+                    '<div data-notify="container" class="col-xl-4 col-lg-4 col-11 col-sm-4 col-md-4 alert alert-{0} alert-with-icon" role="alert">' +
+                    '<button mat-button  type="button" aria-hidden="true" class="close mat-button" data-notify="dismiss">' +
+                    '<i class="material-icons">close</i>' +
+                    '</button>' +
+                    '<i class="material-icons" data-notify="icon">' + icon[color] + '</i> ' +
+                    '<span data-notify="message">{2}</span>' +
+                    '</div>'
+            }
+        );
+    }
+
+    ngAfterViewInit() {
+        this.productDataSource.paginator = this.productPaginator;
+        this.productDataSource.sort = this.productSort;
+
+        this.subProductDataSource.paginator = this.subProductPaginator;
+        this.subProductDataSource.sort = this.subProductSort;
     }
 
     productApplyFilter(filterValue: string) {
@@ -50,8 +83,8 @@ export class ProductsComponent implements OnInit {
     }
 
     ngOnInit() {
-        this.productDataSource.paginator = this.paginator;
-        // this.productsPage.paginator = this.paginator;
+        this.productDataSource.paginator = this.productPaginator;
+        this.subProductDataSource.paginator = this.subProductPaginator;
 
         this.productservice.getProducts()
             .subscribe(
@@ -79,39 +112,67 @@ export class ProductsComponent implements OnInit {
     }
 
     onDeleteProduct(product: Product) {
-        this.productservice.deleteProduct(product.id)
-            .subscribe(
-                data => {
-                    this.productsPage.splice(this.productsPage.indexOf(product), 1);
-                    this.productDataSource.data = this.productsPage;
-                    setTimeout(() => {
-                        this.productDataSource.paginator = this.paginator;
-                    });
-                },
-                err => {
-                    console.log(JSON.parse(err._body).message);
-                }
-            )
-        ;
-        // console.log(client);
+        if (product.subProducts.length > 0) {
+            this.showNotification('bottom', 'right', 0, 'Vous ne pouvais pas supprimé un produit non vide.');
+        } else {
+            this.productservice.deleteProduct(product.id)
+                .subscribe(
+                    data => {
+                        this.productsPage.splice(this.productsPage.indexOf(product), 1);
+                        this.productDataSource.data = this.productsPage;
+                        setTimeout(() => {
+                            this.productDataSource.paginator = this.productPaginator;
+                        });
+                        this.showNotification('bottom', 'right', 1, 'Le produit a été supprimé avec succès.');
+                    },
+                    err => {
+                        this.showNotification('bottom', 'right', 0, 'erreur inconnu.');
+                    }
+                )
+            ;
+            // console.log(client);
+        }
+
     }
 
     onDeleteSubProduct(subProduct: Subproduct) {
-        this.subproductservice.deleteSubProduct(subProduct.id)
+        this.clientservice.getClientsBySubProduct(subProduct.id)
             .subscribe(
-                data => {
-                    this.subProductsPage.splice(this.subProductsPage.indexOf(subProduct), 1);
-                    this.subProductDataSource.data = this.subProductsPage;
-                    setTimeout(() => {
-                        this.subProductDataSource.paginator = this.paginator;
-                    });
+                clientList => {
+
+                    const clients: Array<string> = clientList;
+                    if (clients.length > 0) {
+
+                        let message = 'Le sous-produit que vous voullez suppimé a des client qui y ont accés. [ ';
+                        clients.forEach(client => {
+                            message = message.concat(client + ', ');
+                        });
+                        message = message.slice(0, -2);
+                        message = message.concat(' ]');
+
+                        this.showNotification('bottom', 'right', 0, message);
+                    } else {
+                        this.subproductservice.deleteSubProduct(subProduct.id)
+                            .subscribe(
+                                data => {
+                                    this.subProductsPage.splice(this.subProductsPage.indexOf(subProduct), 1);
+                                    this.subProductDataSource.data = this.subProductsPage;
+                                    setTimeout(() => {
+                                        this.subProductDataSource.paginator = this.subProductPaginator;
+                                    });
+                                    this.showNotification('bottom', 'right', 1, 'Le sous-produit a été supprimé avec succès.');
+                                },
+                                err => {
+                                    this.showNotification('bottom', 'right', 0, 'erreur inconnu.');
+                                }
+                            )
+                        ;
+                    }
                 },
                 err => {
-                    console.log(JSON.parse(err._body).message);
+                    this.showNotification('bottom', 'right', 0, err.error);
                 }
             )
         ;
-        // console.log(client);
     }
-
 }
